requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
selenium==4.12.0
beautifulsoup4==4.12.2
requests==2.31.0
streamlit==1.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.20
pydantic==2.3.0
jinja2==3.1.2
aiofiles==23.2.1
webdriver-manager==4.0.0
psycopg2-binary==2.9.7
sqlmap==1.7.10
plotly==5.16.1
pandas==2.1.0
tqdm==4.66.1
rich==13.5.2
cryptography==41.0.3
pytest==7.4.2
pylint==2.17.5
bandit==1.7.5
colorama==0.4.6
lxml==4.9.3 